Tag the RosBE-Windows 1.0.0 release

+++ reactos/dll/directx/devenum/devenum_private.h
@@ -26,7 +26,6 @@
 
 #include "windef.h"
 #include "winbase.h"
-#include "wingdi.h"
 #include "winuser.h"
 #include "winreg.h"
 #include "winerror.h"
@@ -68,7 +67,7 @@ typedef struct
 typedef struct
 {
     IEnumMonikerVtbl *lpVtbl;
-    LONG ref;
+    DWORD ref;
     DWORD index;
     HKEY hkey;
 } EnumMonikerImpl;
@@ -77,7 +76,7 @@ typedef struct
 {
     IMonikerVtbl *lpVtbl;
 
-    LONG ref;
+    ULONG ref;
     HKEY hkey;
 } MediaCatMoniker;
